.github/actions/oscal-assessment-plan-execute/requirements.txt
PyYAML==6.0
pytest==7.1.1